Hooridor.Ai.AStar

Documentation

data AStar a c

Constructors

AStar

Fields

visited :: !(HashSet a)

waiting :: !(HashPSQ a c ())

score :: !(HashMap a c)

memoHeur :: !(HashMap a c)

cameFrom :: !(HashMap a a)

end :: !(Maybe a)

aStarInit :: (Hashable k, Ord k, Ord c, Num c) => k -> AStar k c

runAStar :: (Hashable a, Ord a, Ord c, Num c) => (a -> HashSet a) -> (a -> a -> c) -> (a -> c) -> (a -> Bool) -> a -> AStar a c

aStar

Arguments

:: (Hashable a, Ord a, Ord c, Num c)

=> (a -> HashSet a)
The graph we are searching through, given as a function from vertices to their neighbours.

-> (a -> a -> c)
Distance function between neighbouring vertices of the graph. This will never be applied to vertices that are not neighbours, so may be undefined on pairs that are not neighbours in the graph.

-> (a -> c)
Heuristic distance to the (nearest) goal. This should never overestimate the distance, or else the path found may not be minimal.

-> (a -> Bool)
The goal, specified as a boolean predicate on vertices.

-> a
The vertex to start searching from.

-> Maybe [a]
An optimal path, if any path exists. This excludes the starting vertex.

This function computes an optimal (minimal distance) path through a graph in a best-first fashion, starting from a given starting point.