Hooridor.Core

Documentation

type Cell = (Int, Int)

data PlayerColor

Constructors

Green

Yellow

Red

Orange

data Inteligence

Who controls the player

Constructors

AI Int
AI algorithm with this depth

Human
Human controls from GUI

data Player

Constructors

Player

Fields

pcolor :: PlayerColor

pos :: Cell

wallsLeft :: Int

inteligence :: Inteligence

data Turn

Constructors

MakeMove Cell

PutWall Wall

type WallPart = (Cell, Cell)

type Wall = (WallPart, WallPart)

data GameState

Constructors

GameState

Fields

playerList :: [Player]

walls :: [Wall]

minRow :: Int

minCol :: Int

maxRow :: Int

maxCol :: Int

size :: Int

cellInBound :: Cell -> Bool

transposeWall :: Wall -> Wall

wallPartEq :: WallPart -> WallPart -> Bool

wallEq :: Wall -> Wall -> Bool

hasWallPart :: WallPart -> Wall -> Bool

areAdjacentCells :: Cell -> Cell -> Bool

isValidWallPart :: WallPart -> Bool

isValidWall :: Wall -> Bool

notInWall :: WallPart -> [Wall] -> Bool

currentPlayer :: GameState -> Player

oneStep :: Cell -> Cell -> [Wall] -> Bool

takeTurn :: Turn -> GameState -> GameState

Perform one turn and give turn order to next player

validTurn :: Turn -> GameState -> Bool

Checks if Turn is valid

move :: GameState -> Turn -> GameState

Generate new state from movement, but don't pass turn

validMoves :: GameState -> [Turn]

Get all possible moves

isWinner :: Player -> Bool

playersCanReachGoal :: GameState -> Bool

Check if all players can reach final state

playerCanReachGoal :: GameState -> Player -> Bool

Check if player can reach final state

playerCanReachGoalOld :: GameState -> Player -> Bool

giveTurn :: GameState -> Player -> GameState

Give turn to some player

availablePositions :: GameState -> Player -> Cell -> [Cell]

In this GameState this Player standing on this Cell | to what cells can go in one step

putPlayerOn :: GameState -> Player -> Cell -> GameState

Make a state where it is this player on given cell an it's his/her turn to move

getShortestPath :: GameState -> Maybe [GameState]

Get shortest path to victory of current player

winner :: GameState -> Maybe Player

Winner in this game state, if there is one.

defaultWalls :: Int

initialStateAi :: Int -> GameState

Initial state for a 2-player game, where the second is an AI algorithm with a given depth.

initialState :: Int -> GameState