Hooridor.Gui

Documentation

data GuiState

State of the GUI: current GameState and additional decoration to draw.

Constructors

GuiState GameState Picture

data BoardObject

Board objects that can be pointed at by a mouse.

Constructors

BoardCell Cell

BoardWall Wall

window :: Display

Default Display to draw the game in.

background :: Color

Default window background color.

fps :: Int

Default game frames per second value.

normalCellColor :: Color

Display color of normal cells

normalWallColor :: Color

Display color of walls

colorPlayer :: PlayerColor -> Color

Display color of PlayerColors

cellSize :: Int

Side of square cells in pixels.

playerRadius :: Float

Radius of the player icon

cellMargin :: Int

Free space between cells in pixels.

cellSizeAndMargin :: Int

Distance between sides of two adjacent cells.

Equal to cellSize plus cellMargin: prop> cellSize + cellMargin = cellSizeAndMargin

cellCenter :: Cell -> Point

Center of the Cell drawn on the screen.

wallPartCenter :: WallPart -> Point

WallPart center coordinates.

pointingAt :: Point -> Maybe BoardObject

What board object is drawn at this point.

handleEvents :: Event -> GuiState -> GuiState

Response to events.

playerIcon :: Player -> Picture

Player icon, not translated.

drawPlayer :: Player -> Picture

Player icon translated to the cell position.

drawCell :: Cell -> Color -> Picture

Colored cell translated to the cell position.

drawDefaultCell :: Cell -> Picture

Cell with the default color translated to the cell position.

highlightCell :: Cell -> Player -> Picture

Cell highlighted for this player and translated to the cell position.

meanPoint :: Point -> Point -> Point

The point between two points.

drawWallPart :: WallPart -> Color -> Picture

Draw a wall part with this color, translated to the correct position.

drawWall :: Color -> Wall -> Picture

Draw wall with this cololr, translated to the correct position.

wallHint :: Wall -> Player -> Picture

Draw a wall hint, translated to the correct position.

drawBoard :: Picture

Draw all cells normal, translated to the correct positions.

drawVictoryScreen :: Player -> Picture

Draw victory screen for this player.

currentPlayerIndicator :: Player -> Picture

Picture telling how many walls this player has, translated on top of the board.

render :: GuiState -> Picture

Draw current game state

initiateGame :: GMMenuResult -> GuiState

Create new GuiState with this number of players.

update :: Float -> GuiState -> GuiState

Update per time (nothing updates).

runGui :: IO ()

Run the game starting with the menus.